server/requirement.txt
flask==2.3.2
flask-cors
mysql-connector-python
flask-jwt-extended
bcrypt==4.0.1